UI Clean up

+++ resources/js/Pages/PostsList.vue
@@ -1,5 +1,6 @@
 <script setup>
 import {router} from '@inertiajs/vue3';
+
 defineProps({
     posts: Array
 })
@@ -8,43 +9,47 @@ const deletePostById = (postId) => {
     router.delete(`/posts/${postId}`)
 }
 const handleDeleteButtonClicked = (postId) => {
-    deletePostById(postId)
+    if (confirm('Delete post?') === true) {
+        deletePostById(postId)
+    }
 }
 </script>
 <template>
     <div>
         <div>
-            <h1>Posts</h1>
+            <h1 class="mb-4 text-4xl font-extrabold leading-none tracking-tight text-gray-900 md:text-5xl lg:text-6xl">
+                Posts
+            </h1>
         </div>
-        <div>
-            <table>
-                <thead>
-                <tr>
-                    <td>ID</td>
-                    <td>Title</td>
-                    <td>Username</td>
-                    <td></td>
-                    <td></td>
-                </tr>
+        <div class="relative overflow-x-auto">
+            <table class="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
+                <thead class="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
+                    <tr>
+                        <th scope="col" class="px-6 py-3">ID</th>
+                        <th scope="col" class="px-6 py-3">Title</th>
+                        <th scope="col" class="px-6 py-3">Username</th>
+                        <th scope="col" class="px-6 py-3"></th>
+                        <th scope="col" class="px-6 py-3"></th>
+                    </tr>
                 </thead>
                 <tbody>
-                <tr v-for="post in posts">
-                    <td>
-                        <span>{{post.id}}</span>
-                    </td>
-                    <td>
-                        <span>{{post.title}}</span>
-                    </td>
-                    <td>
-                        <span>{{post.user.username}}</span>
-                    </td>
-                    <td>
-                        <a :href="'/posts/' + post.id">View</a>
-                    </td>
-                    <td>
-                        <button v-on:click="handleDeleteButtonClicked(post.id)">Delete</button>
-                    </td>
-                </tr>
+                    <tr v-for="post in posts" class="bg-white border-b dark:bg-gray-800 dark:border-gray-700">
+                        <td class="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                            <span>{{ post.id }}</span>
+                        </td>
+                        <td class="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                            <span>{{ post.title }}</span>
+                        </td>
+                        <td class="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                            <span>{{ post.user.username }}</span>
+                        </td>
+                        <td class="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                            <a :href="'/posts/' + post.id">View</a>
+                        </td>
+                        <td class="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                            <button v-on:click="handleDeleteButtonClicked(post.id)">Delete</button>
+                        </td>
+                    </tr>
                 </tbody>
             </table>
         </div>
